Use Adapters to create both pandas and polars containers
